book: translation of ch05
Update #12
