+++ book/zh-cn/toc.md
@@ -68,6 +68,7 @@
     + 5.1 RAII 与引用计数
     + 5.2 `std::shared_ptr`
     + 5.3 `std::unique_ptr`
+    + 5.4 `std::weak_ptr`
 - [**第 6 章 标准库: 正则表达式**](./06-regex.md)
     + 6.1 正则表达式简介
         + 普通字符
