+++ code/5/5.3.weak.ptr.cpp
@@ -1,3 +1,13 @@
+//
+// 5.3.weak.ptr.cpp
+// chapter 05 start pointers and memory management
+// modern c++ tutorial
+//
+// created by changkun at changkun.de
+// https://github.com/changkun/modern-cpp-tutorial
+//
+
+
 #include <iostream>
 #include <memory>
 
@@ -8,14 +18,14 @@ class A {
 public:
     std::shared_ptr<B> pointer;
     ~A() {
-        std::cout << "A 被销毁" << std::endl;
+        std::cout << "A was destroied" << std::endl;
     }
 };
 class B {
 public:
     std::shared_ptr<A> pointer;
     ~B() {
-        std::cout << "B 被销毁" << std::endl;
+        std::cout << "B was destroied" << std::endl;
     }
 };
 int main() {
